backend/requirements.txt
openpyxl>=3.1.0
python-docx>=1.0.0
nbformat>=5.9.0
sqlglot>=20.0.0
pymupdf4llm>=0.0.5
python-pptx>=0.6.21
